requirements.txt
asgiref==3.11.0
cmdstanpy==1.3.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
cycler==0.12.1
Django==5.1.3
django-crispy-forms==2.5
et_xmlfile==2.0.0
fonttools==4.60.1
holidays==0.85
importlib_resources==6.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
narwhals==2.12.0
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
prophet==1.2.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
stanio==0.5.1
tqdm==4.67.1
tzdata==2025.2
